[split] Futures: simplify cancellation, and guarantee closure invocation order.
cancellation:

  - the semantics of Futures themselves remain unchanged.  we're back
  to one source of updates, and there are no special cancellation
  exceptions - instead, a Future becomes a Cancellable.  Cancellables
  simply propagate a level-triggered signal (with linkTo) and allow
  interested receivers to act on it

thus: cancellation is now a signal, and *if* the provider of a future
is interested, it will do whatever is an appropriate action.  this
means that the semantics of cancellation is wholly within the provider
of the value, and this is anyway where the value semantics lie (eg.
can this fail?, what are the kinds of exceptions, etc.).

change the definition of closure invocation ordering: we no longer
provide order-by-respond, but rather order-by-chaining.  it turns out,
however, that the simplest implementation (that i can think of) of
order-by-chaining is actually order-by-respond, so we implement this
(via a new concurrency primitive; IVars).