util-logging: Make Level objects thread-safe
The Level objects and the Logger object are intertwined in a
way that may lead to trouble in multithreaded applications: the
Level class constructor requires access to the Logger object,
whose constructor, in turn, forces construction of the Level
objects. Because evaluating the constructor of an object blocks
other threads trying to dereference the object, this can end up
in a deadlock:

  - Thread 1 holds lock for 'Level.OFF' construction and awaits
    on the Logger object
  - Thread 2 holds lock for the Logger object construction and
    awaits on 'Level.OFF'

Fix the problem by constructing the Level objects non-lazily when
the Level object (rather than the Logger object) is accessed for
the first time.