[split] UpdatableVar: fine-grained locking to prevent deadlocks.
Deadlocks may happen because observation and updating with respect to
derived Vars happen in the opposite order, and some operations (such
as observeUntil) manipulate state synchronously with updates.

A simple example:

	val a = Var(333)
	val b = Var(a)
	val c = b flatMap identity

	thread 1:
		b.observeUntil(_==333)
	thread 2:
		a() = 333

here, thread 1 is racing with thread 2: t1 may observe, then
synchronously close(); however in the meantime t2 may try to lock a,
then, b, then, c whereas t1 already holds the lock for b while it's
trying to close a. This leaves us in a situation where t2 holds the
lock on a, t1 holds the lock on b while t2 is trying to get the lock
for b, and vice-versa.

Another fix for this would be to have observeUntil participate in
future scheduling, but that seems like a hack.

RB_ID=245305