[split] Var: fix an iatrogenic concurrency bug
The current observe code is:

    val o = O(obs, depth, version.getAndIncrement())
    val newOwner = obs.lease(this, o.version)

    if (newOwner) {
      obs.update(state.get.v, this)
      cas(_ + o)
      ..

it was possible here to miss updates between updating the observer for
the first time, and CASing in the O representation, leaving the
observer with a stale value.

The fatal flaw here is that updating observers has to be atomic with
publishing values. (This really violates the posited single-owner
policy.)

I took this opportunity to simplify the code for Leases, pushing in
most of the single-owner responsibility there, and reworking the API a
bit. The result is that the code is much simpler to reason about: we
have strong guarantees around atomic state transitions, with the
ability to resolve races by maintaining value versions.

RB_ID=268363