[split] util-core/Offer: Don't do indexed lookups in prepare()
Motivation:

Colony was showing a large CPU bottle neck, which was profiled to the
Offer sync() method, which in turn calls prepare()/choose().  The
prepare method mapped the input Seq to an internal array, however it
depended on calling .size() and indexed lookup on the Seq on each
iteration.

Modifications:

This patch keeps the same style (avoiding inner closures and temporaries),
but switches to caching the size and using an iterator, instead of
indexed lookup.

Result:

This change brought Colony's ZK index startup time down to 2
minutes (network latency bound) from 40+m (which was CPU bound).

RB_ID=296321