[split] Implement application-level shutdown handling in App.
Motivations:
App currently implements onExit using Runtime.addShutdownHook. This comes with all sorts of baggage in terms of what can actually occur in a shutdown handler. For instance, logging and network resources cannot be relied upon in this context. In particular, this makes things like flushing a queue of work tricky.

Modifications:
Make App extend Closable so that it may be closed with a deadline.  ShutdownHandler now takes a Closable that is called before calling System.exit. The admin endpoints /quitquitquit & /admin/shutdown now take a 'grace' query parameter which can be used to control how long to give the App before calling System.exit. App.onExit no longer registers shutdown hooks via Runtime.addShutdownHook and instead manages a list of Closable "exits" which are closed when the app is closed.

Results:
When the application is shut down gracefully, onExit and postmain hooks will not run in the context of a VM shutdown hook; and they may therefore utilize, for example, logging and networked resources.

RB_ID=285287