[split] Make com.twitter.finagle.Name an ADT
Problem

There is currently a proliferation of Name implementations which are
related in all sorts of convoluted ways. This is accidental complexity
due to the fact that the Name interface, being part of Finagle's
public API, remains sacrosanct -- at least for the moment.

Solution

This change tries to simplify things a little by making Name a proper
ADT with two variants: Name.Bound describes a bound address; Name.Free
is an unbound Name whose resolution is determined by a Namer. This
makes the relationships clear, and requires only a small API change.

It also makes clear that, in intermediate computation, there are two
levels of dynamism: the tree itself is dynamic; its nodes are
independently dynamic. There is still confusion between the
relationship of Var.{Neg, Pending, etc..} with that of
Activity.Ok(NameTree.{Neg, Empty}) | Activity.Pending, etc. This is
unfortunate -- ideally, we'd be able to pull all of that out of Addr,
and model it entirely as NameTrees. But this is about as far we can go
without making major API changes.

It also has a simplifying effect on Namer signatures. Namely, binding
is now more explicit:

	def lookup(path: Path): Activity[NameTree[Name]]
	def bind(tree: NameTree[Path]): Activity[NameTree[Name.Bound]]

and it is clear whence dynamism is derived.

I also took the opportunity to unify some of the naming (heh.) around this.
In particular, InterpretingFactory is now BindingFactory.

Result

Nicer API; no functionality changes.

RB_ID=330105
TBR=true