[split] util: Add a new sbt target to publish locally to m2
Problem

Publishing the maven-scrooge-plugin requires having access to the latest util.
Every time we publish, we need to wait for the util release to be available on Sonatype.

Solution

Rather than waiting for the release to be available on Sonatype, this new sbt target
installs it locally.

Result

./sbt +publish-m2 publish the jars in the local maven repository

RB_ID=337145