[split] util-core: Introduce ConcatBuf
Problem

The old implementation of `Buf.concat` was returning an anonymous implementation of Buf as a
deeply nested tree. Any call to one of the methods of this Buf was potentially causing a
StackOverflowError

Solution

Introduce a new class ConcatBuf, which represents a linked list of Buf instead.

Result

Every call to a method of ConcatBuf is now iterating through the link list, which doesn't
consume the memory in the call stack.

RB_ID=362791
TBR=true