[split] Add additional handlers to Logging trait.
Problem

There is no way to override the handlers that the default LoggerFactory uses in the Logging trait. The result is that if someone wants to add a handler to the root logger factory (e.g. for LogLens) they need to override and copy paste the whole method.

Solution

Add a method called handlers that can be overridden if needed. This initiates one handler by default, but it can be overridden if we need to add more handlers.

Result

It is now possible to easily configure LogLens for all applications.

RB_ID=375114