[split] util-core: Fix update-after-interrupt race condition in AsyncSemaphore
Problem

Production logs from a rate-limiting service indicated that a Promise
is being double-set somewhere within
RequestSemaphoreFilter/AsyncSemaphore.

Solution

We think that the issue is being caused by a race between invocations
of the interrupt handler set in `AsyncSemaphore.acquire` and the
unsynchronized `Promise.setValue` in `AsyncSemaphore.release`. With
this change, `release` is synchronized in its entirety, rather than
just the queue-poll.

RB_ID=387575