[split] util-core: Catch InterruptedException in Closable collector thread
Problem

A report on the finaglers mailing list indicated that the ThriftClient
class in finagle-example fails to terminate cleanly when run in SBT [1].
I think this has to do with SBT's handling of System.exit, explained in [2].

> "User code is run with a custom SecurityManager that throws a custom
> SecurityException when System.exit is called. This exception is caught by sbt.
> sbt then disposes of all top-level windows, interrupts (not stops) all
> user-created threads, and handles the exit code."

Thus even though daemon threads should be killed outright on process
termination, SBT instead interrupts them. This causes the Closable
collector thread (which is blocked on `ReferenceQueue.remove()`) to
log and throw a fatal exception.

Solution

Add a pass-through case for InterruptedException within the collector
thread.

[1] https://groups.google.com/forum/#!topic/finaglers/Hq1CHqeeLgI
[2] http://www.scala-sbt.org/release/docs/Detailed-Topics/Running-Project-Code.html

RB_ID=278895