Add redirects from old API docs to new
Problem

We're still hosting old versions of the API docs (from before Unidoc). These
out-of-date versions are appearing before the current version in search results.

Solution

I've written a small script to add the appropriate meta redirect and canonical
link to each page in the older versions (which are not touched by the updatedocs
script).

Result

I ran the automatically generated links through a little PhantomJS test, and the
following are broken:

https://twitter.github.io/util/docs/#com.twitter.concurrent.Channel
https://twitter.github.io/util/docs/#com.twitter.concurrent.ChannelSource
https://twitter.github.io/util/docs/#com.twitter.concurrent.Observer
https://twitter.github.io/util/docs/#com.twitter.concurrent.ObserverSource
https://twitter.github.io/util/docs/#com.twitter.logging.BasicFormatter$
https://twitter.github.io/util/docs/#com.twitter.logging.config.BasicFormatterConfig$
https://twitter.github.io/util/docs/#com.twitter.util.Eval$$CompilerException
https://twitter.github.io/util/docs/#com.twitter.util.Eval$$StringCompiler
https://twitter.github.io/util/docs/#com.twitter.util.Locals$
https://twitter.github.io/util/docs/#com.twitter.util.MapMaker$
https://twitter.github.io/util/docs/#com.twitter.util.MapMaker$$Config
https://twitter.github.io/util/docs/#com.twitter.util.SavedLocal
https://twitter.github.io/util/docs/#com.twitter.util.SavedLocals
https://twitter.github.io/util/docs/#com.twitter.util.StreamHelper$
https://twitter.github.io/util/docs/#com.twitter.util.TimeMath$
https://twitter.github.io/util/docs/#com.twitter.util.TimeOverflowException
https://twitter.github.io/util/docs/#com.twitter.util.TryLike

With the exception of CompilerException, these actually don't exist now, so I
think it makes sense simply to leave the links broken.