[split] util-doc: Fix updatedocs.bash to update new util docs
Problem

I noticed that we weren't actually writing new util docs.  We only
run ./sbt unidoc, which generates docs for 2.10 (the current default
scala for the sbt project) but not for 2.9.2, which the script had
previous been hardcoded to copy over.  Although `set -e` had been
set, there was a cached 2.9.2/unidoc which it could copy over, so
we never detected a problem.

Solution

rm -rf the old unidoc directory, and switch over to using 2.10.

RB_ID=424303