[split] Copy bytes from Reader to Writer and OutputStream
Problem

The initial problem was to write all the bytes from a Reader into an
OutputStream. It was suggested that a nice way to do this is to have a
general way to copy the byte stream from Reader to Writer, and a way to
give OutputStream a Writer interface.

Solution

This changeset defines Reader.copy for copying byte streams from
Readers into Writers; and Writer.fromOutputStream for turning
OutputStreams into Writers.

Result

A convenient way to write bytes from a Reader to a Writer, and by
extension, from a Reader to an OutputStream.

RB_ID=413781