[split] util-core: Changes transform to fail Futures if you return inside the passed closure
Problem

return statements in closures are implemented by turning the return
statement into `throw NonLocalReturnControl` and then unpacking the
contents of NonLocalReturnControl at the top level of the method.
If the closure isn't executed inside of the method, for example if
the execution is deferred, it gets screwed up.  This can be difficult
to debug, especially since the Future is never returned.

Solution

A different Future is only returned by transform, so I chased down
where transform invokes the closure, and caught the NonLocalControlReturn
(NLCR) exceptions explicitly.  In scala's implementation, they
fulfil the future with the result from the NLCR, but that seems
totally crazytown, because it will almost certainly be the wrong
type.  Instead, we wrap the NLCR in a NonLocalWrapper, and fail the
Future.

https://github.com/scala/scala/blob/2.11.x/src/library/scala/concurrent/impl/Promise.scala#L52

Result

If you stick a return in a closure and pass it to a Future, you'll
get a nice failed future which explains why your future has been
failed.

RB_ID=429801