[split] Fix flaky Reader.copy test
The test asserted the equality of two OutputStreams by by converting
them into byte arrays then testing the byte arrays. It may be the case
in some circumstances, that the conversion to byte array requires a
flush to ensure that all the data written to the OutputStream is
contained in the resulting byte array. This change calls flush on both
OutputStreams before converting to byte arrays.

RB_ID=435699