[split] util-core: Make register/close on Event() work atomically
Problem

Event[T]() creates an instance that maintains a list of registered
witnesses in @volatile var registerrs. The updates to this var are neither
atomic nor synchronized. As a result adding/closing a witness is not
guaranteed to execute successfully.

Solution

Wrap every access of the witnesses list with synchronized block.
+Use a Vector instead of a List for efficiency.

This Commit is comming from github.com/sirmax, partially modified by me.
Signed-off-by: Steve Gury <stevegury@twitter.com>

RB_ID=382203