[split] util-app: introduce Flag.let
Problem

Due to `com.twitter.app.Flag`s being assigned values via the
command-line or System properties, it is difficult to write
tests that require different values for the Flags.

Solution

Introduce `Flag.let(t: T)(f: => Unit): Unit` which overrides
the value of the flag with the value `t` for the function `f`
for the scope of the current `com.twitter.util.Local`.

Result

Previously difficult to test code is now testable.

RB_ID=456653