[split] c.t.util.Time: Scope Locals with `Local.let` instead of `save`+`restore`
Problem

`Time.withTimeFunction` currently saves and restores the entire Local
array as a means of scoping a time function and timer. This hasn't led
to problems per se, but it's a heavy-handed approach when the ability
exists to scope individual local items instead of the entire set.

Solution

Use `Local.let` instead of `Local.save` and `Local.restore` to scope
the Time-related Local fields.

Result

Uses of `withCurrentTimeFrozen` and `withTimeAt` in tests will
selectively scope the Time-related Local fields instead of blowing
away the entire Local array.

RB_ID=462029