[split] util-core: prefer Await.result(future.liftToTry) to deprecated methods
Problem

It is not obvious to clients what the preferred mechanism is for
getting back the result of a `Future` as a `Try`.

Solution

Add documentation to `Await` that leads people to use a non-deprecated
approach of `Await.result(future.liftToTry)`.

Update existing code to use this mechanism as well to avoid the
deprecated `Future.get(Duration)` call.

Result

Less usage of deprecated code, better docs and make this
approach more obvious for future coders.

RB_ID=463483