[split] util-core: Perf improvement to ConcatBuf#slice
Problem

ConcatBuf's slice method was calling concat for every Buf in the
new slice, and concat allocates a new Vector object for every
concatenation.

Solution

Since we know that we only need to slice the first and last item
of the new slice, and the interstitial ones will be unchanged, we
can get this method down to just three necessary allocations, the
new ConcatBuf, two for prepending and appending the first and last
ends of the slice.

Result

ConcatBuf#slice is constant in its GC pressure rather than linear.

old:

0% Scenario{vm=java, trial=0, benchmark=SimpleSlice} 81,898,250.00 ns; ??=809092.52 ns @ 6 trials, allocated 10,567,956 instances for a total of 394,485,000B
50% Scenario{vm=java, trial=0, benchmark=ComplexSlice} 1,161,000,000.00 ns; ??=320684887.56 ns @ 10 trials, allocated 110,692,536 instances for a total of 4,895,614,920B

new:

0% Scenario{vm=java, trial=0, benchmark=SimpleSlice} 81,533,980.77 ns; ??=7463158.54 ns @ 10 trials, allocated 5,480,854 instances for a total of 340,843,200B
50% Scenario{vm=java, trial=0, benchmark=ComplexSlice} 414,192,750.00 ns; ??=11997791.36 ns @ 10 trials, allocated 8,300,670 instances for a total of 675,817,536B

RB_ID=461809