[split] util.RingBuffer: fix buffer size on drops
Problem
util.RingBuffer has a `drop` method that removes several elements from the buffer at a time. However, it doesn't reduce the internal size, which is expected. This operation does reset the size to 0 if we're removing all elements, but not if we drop fewer than all elements.

Solution
Decrease the count_ (internal size) when drop is performed
Also: deprecate it in favor of Guava (see review)

RB_ID=480321