[split] removed IVar and IVarField Problem
IVars originally backed the twitter Future implementation, but the IVar behavior was inlined when we reimplemented Futures. Now, IVars support a subset of Promise behavior and are rarely used. We found that they were introducing a cyclic dependency between scala and java code, which was fragile and breaking some internal tools.

Solution

Remove IVar and IVarField

Result

IVar won't be available.

RB_ID=477763