[split] Add asynchronous cache with TTL to util-cache
Problem

Services sometime provide endpoints that don't change very frequently,
for example the trend service's list of supported locations.  In some
cases these responses are quite large so we want to minimize how often
the data is requested.  We need a lightweight way to cache successful
responses of these methods while retrying on failures that allows a TTL
so results are periodically refreshed.

Solution

This change adds 'Refresh' which takes a TTL and function that returns a
future.  The first call to the memoized future calls through to provider
function, but subsequent calls will just get the cached copy until the
TTL expires, after which it will be requested once.  Special care has
been taken to handle race conditions on requests.

Result

After this change, clients will be able to wrap any function that
returns a future with this utility that caches responses for up to the
specified TTL.

RB_ID=480295