[split] finagle/util - Make multithreaded tests more patient
Problem

By default, the `eventually` method in scalatest times the test out
after 150 milliseconds.  This makes some of our unit tests appear flaky
when random pauses or busy CI machines stall on test execution.  The
symptom of these types of failures look like this:

       org.scalatest.exceptions.TestFailedDueToTimeoutException: The
       code passed to eventually never returned normally. Attempted 1
       times over 159.628854 milliseconds.

The default patience config for scalatest[1] should try every 15ms, so
with only one retry it's clear that the CI machine is slowed for some
reason.

Solution

Mix in the IntegrationPatience trait which increases the timeout of the
`eventually` to 15 seconds.

1. http://www.artima.com/docs-scalatest-2.0.M5/org/scalatest/concurrent/PatienceConfiguration.html

RB_ID=504031