[split] util: Local.let Problem:
A common pattern in code that needs to save and restore Local contexts is to save in one place, then later swap the current with the saved, do some work, then swap back.

Solution:

Add Local.let, which takes a previously saved context, swaps it for the current context, executes a call-by-name function, the swaps back.  This can remove some boiler-plate code and removes the risk of forgetting to reset the context via try/finally.

RB_ID=514603