[split] util-app: GlobalFlags to load only Scala object classes
Problem:

`GlobalFlags` has a discovery mechanism to find any global flag classes,
where the classpath is scanned and every class is loaded to check if it is
annotated with `GlobalFlagVisible` and if it is a Scala object class.

This discovery mechanism is an expensive operation, and it might fail if
classpath contains classes with incompatible definitions. This scenario might
happen if classpath contains different versions of the same library (e.g.
application using Guava library in a Hadoop environment).

Solution:

Instead of trying to load any class, restrict it to only Scala object class.
it will speed up the discovery, and avoid issues regarding class loading.

RB_ID=513545