util-logging: now usable in java!
Problem

Logger is painful to use in java due to all of the logging methods being varargs.
Also there's a bug in the LoggerFactoryBuilder shim that could do a better job
being type-compatible with various logging handlers.

Solution

Add the @varargs annotation to the logging methods and expand the type scope of
the LoggerFactoryBuilder.addHandler method.

RB_ID=510781