Use master branch status only for Travis CI badges
Problem

Without a branch parameter in the image URL, the Travis CI badge will show the
status of the latest build. We don't guarantee that develop branches will
compile without local publication of dependencies, which means the badge may not
always be accurate.

Solution

Add a branch parameter so that we only show the build status of master.

RB_ID=544318