util-core: Evaluate Closable.{all, sequence} eagerly; don't Await in Var.flatMap
Problem

Var.flatMap calls Await.ready in order to ensure that the effects of
Closable.close() have propagated. (Var currently defines propagation
to be synchronous.)

This can cause deadlocks because the Var's synchronization assumes it
isn't possible to interleave updates on a single thread. However,
Await.ready can introduce this as it flushes its schedule (in order to
prevent a deadlock where a future is updated in an activation that is
now left behind).

Solution

Ensure that we don't call Await.ready in Var.flatMap.

Note

I'm not happy about this solution, but it's a reasonable workaround.
I'd really like to change Var's propagation semantics to be
asynchronous, to avoid this class of problems altogether. It's
difficult to avoid synchronized updates when implementing synchronous
propagation while also wishing to make sure no updates are missed.

RB_ID=557464