Easy argument capture for Mockito mocks.
Mockito's ArgumentCaptor enables you to capture the arguments that were
passed to stubbed-out methods on mocks, to make assertions on them.
However, the API for ArgumentCaptor was built for Java and is kind of
clunky compared to what we expect in Scala.

An alternative to using ArgumentCaptor is to use Mockito argument
matchers in `when(myObj.method(...))`. However, this means every
assertion you want to make has to be encapsulated in a custom matcher
subclass, and you can't make assertions spanning more than one argument
in each stubbed method invocation. For example, you might want to check
that a key-value store wrapper object was called with a key of "foo" and
a timeout of 10 seconds, and Mockito argument matchers would not be able
to provide this.

Another alternative is to use matchers in a
`verify(myMock, times(N)).someMethod(...)` call, but this suffers from
the same problem of having to write a specific matcher subclass for any
non-standard assertions.

The ArgumentCapture trait can be mixed into your test class and has two
methods, `capturingOne` and `capturingAll`, which capture the
invocations of your stubbed methods as a tuple or Seq of tuples. Here's
an example:

    when(myObj.lookup(any[String], any[Duration])).thenReturn(someValue)

    // The call to myObj.lookup will normally live somewhere inside the
    // subject under test.
    assert(myObj.lookup("foo", 10.seconds) === someValue)

    val (key, timeout) = capturingOne(verify(myObj).lookup _)
    assert(key === "foo")
    assert(timeout === 10.seconds)

An example which uses `capturingAll` and doesn't make any assertions
about the order of the stubbed method calls (`forExactly` is defined on
the `org.scalatest.Inspectors` trait):

    val requests = capturingAll(verify(myObj, times(2)).lookup _)

    forExactly(1, requests) { (key, timeout) =>
      assert(key === "foo")
      assert(timeout === 10.seconds)
    }

    forExactly(1, requests) { (key, timeout) =>
      assert(key === "bar")
      assert(timeout === 2.seconds)
    }

The only issue with this solution is that it currently only has
implementations up to 5-ary functions; vanilla Scala doesn't have
variadic type parameters so the captureOne/captureAll functions for each
arity have to be copied, pasted and extended by hand currently.

Rather than writing stricter `when(myObj.method(...))` matchers, you can
write arbitrarily loose stub pattern matchers, capture one or all of the
invocations, and make assertions on them.