[util-collection] default field values should be evaluated at most once per record instance
Problem
-------

Field[A] stores its default value as a val. If the type `A` is mutable, however, it's not safe to share this reference. For example:

  // declare a mutable data type, and a field with a default value of that type.
  class MutableInt(var value: Int)
  val schema = new RecordSchema
  val field = schema.newField(new MutableInt(1))

  // field is not set in newRecord, so get the default. then, mutate it.
  val first = schema.newRecord.apply(field)
  first.value = 2

  // field is again not set in this newRecord, so get the default and suffer.
  val second = schema.newRecord.apply(field)
  assert(second.value == 1) // fails, because `first eq second`.

Solution
--------

Default values for fields should be reevaluated on demand for each new record. Once a default is evaluated for a particular record, it should be stored in that record as though it had been set explicitly. Semantics now match those of Scala's built-in method/constructor parameter defaults.

Result
------

It is now safe to declare a field with a default value for mutable data types.

RB_ID=573392