Add Try.throwable for friendlier Java access to failure conditions.
Motivation:

It is currently non-ergonomic to extract the failure (exception) from
a Future from Java in circumstances where you want to treat the
failure by value rather than propagate as an exception. For example, a
unit test wanting to confirm the type (or other detail) of a failure
signaled by an asynchronous computation.

Solution:

Augment Try to have a throwable method returning Throwable,
such that liftToTry() can be used:

  final Try<Throwable> try = Await.result(aFuture.liftToTry());
  if (try.isThrow()) {
    final Throwable t = try.throwable()
  } else {
    ...
  }

Try.throwable throws an IllegalStateException in the event that it is
called on a Return.

Result:

The change is additive and should not impact existing code. It the
original problem, and also seems to be a legitimate augmentation of
Try itself independently of it. Given that Try represents the result
of a computation in a way that preserves error conditions, it makes
sense to expose such error conditions in a friendly way to callers.

Prior to the change, this was not always the case for Java callers.

RB_ID=583490