util-registry: Naive implementation to get the ball rolling
Problem

We want to have a low-level API for the R* Registry, which codifies
the rules that we expect the "client libraries" that we actually
ship to people will use.

Solution

Nail down the API with a naive implementation first.  The naive
implementation is partially there because the data model is exactly
what I imagine it being in my head.

RB_ID=568274