util-collection: Add copy to RecordSchema.Record.
Problem

It is impossible to make a copy of a RecordSchema.Record.  This means that it is
impossible to copy objects which have a Record as a member such as
com.twitter.macaw.test.TestRequest.

Solution

Implement a copy method on RecordSchema.Record.

Result

It is now possible to copy RecordSchema.Records.  Work on TestRequest is
unblocked.

RB_ID=590416