util-registry: Expose a public interface for libraries to register themselves
Problem

Other libraries should be able to export labels too, but we want
to let them do it in a mostly structured way

Solution

Provide a Library.register endpoint which only lets you register
in a small namespace, and disallows reregistering with the same
name.  It also exposes an api for changing existing params.

RB_ID=593302