Fix {Queueing,Throttled}Handler method call delegation issue
Problem

QueueingHandler and ThrottledHandler are wrappers around another handler. But not all Handler methods are correctly delegated to the wrapped instance.

TwitterServer wraps FileHandler into a QueueingHandler, and just after, calls setFormatter(), but the call is never propagated to FileHandler. As a consequence, records are formatted with the default formatter.

Solution

Introduces a ProxyHandler abstract class to delegate all methods to the wrapped handler, and retrofit QueueingHandler and ThrottledHandler to use it

Result

All Handler method calls are forwarded to the wrapped instance.

RB_ID=600938