Change class-to-trait inheritance to self type
Problem

The behavior of traits that extend classes in Scala seems to be underspecified
and buggy, and this code breaks between 2.11.4 and 2.11.5 (see my new SI-9239).

Solution

Switching to a self type works on 2.11.5, isn't flagged by MiMa as an ABI
change, and can't break code outside of Util because the trait is sealed.

Result

We can upgrade our 2.11 build to 2.11.6.

RB_ID=613760