Replace reference to scala.ScalaObject with scala.AnyVal
Problem

ScalaObject was deprecated in Scala 2.10, so referencing
"scala.ScalaObject" to determine the Scala library path does not
work in Scala 2.11 and onwards.

Solution

Replace the reference to "scala.ScalaObject" with a reference
to "scala.AnyVal", which serves as the root class of all value
types, describing all values not implemented as objects in the
underlying host system.

This resides in the same jar as ScalaObject--and is also less
likely to be deprecated in future Scala versions--so it will
provide the same class path.

Result

`classPathOfClass("scala.ScalaObject")` will be replaced with
`classPathOfClass("scala.AnyVal")`.