finagle-core: eliminate some allocations in context processing
Problem

The deadline branch introduced a standard context that is now shipped
in basically all clients and servers. This resulted in a relatively steep
allocation regression in the mux smoke tests (~20%). With this change,
we recoup all of it.

Solution

Reduce allocation during key processing and comparison:

1. Make sure context keys are ByteBuffers so that we can;
2. compare them directly without copying them.

On master, we get this in the Mux allocation test

	before 8552 bytes/iter
	after 18486 bytes/iter

The effect is (much) more pronounced in the deadlines branch.

RB_ID=606326
TBR=true