birdcage: Remove empty util-cache/java target
java_library targets without sources are basically target aliases (`target(name='..', ..)`), which don't have provides clauses.

I suspect that you meant to include some sources in this target? But if you didn't it shouldn't be used, because it would create false dependencies between the libraries that it depends on.

RB_ID=626970
TBR=true