[trivial] Event sink default should match root logger level
Problem

EventSink defaults to capturing the root logger at Level.ALL.  This is
problematic because some log messages are computed lazily, and it would
be too easy to unwittingly force these messages with EventSink's
defaults.

Solution

Match the log level of the root logger. This is what service owners
would set to a reasonable default.

RB_ID=639266