[util-core] Refine AsyncStream.flatten
Problem

It's too hard to flatten a stream of streams. AsyncStream.flatten
doesn't do it. We want something that works like Seq.flatten.

Solution

Introduce a flatten method on the stream that works like Seq.flatten, in
that it concatenates all streams in a stream.

We also remove AsyncStream.flatten since it can be recovered by
composing the new flatten and fromFuture.

    AsyncStream.flatten(s) = AsyncStream.fromFuture(s).flatten

RB_ID=657264