util-core: fix credentials parsing of values containing `?`
Problem:

RB 679812 introduced an exception when constructing a Credentials with a value containing a `?`.

Solution:

Fix the linebreak regex to not include `?` (I previously overlooked that it was in a `[]`).  For the picky: note that this change will allow bare `\r`s as line breaks, while the previous change only allowed `\r\n` and `\n`, but the yaml spec actually says those are legal line breaks due to MacOS <= 9 (http://www.yaml.org/spec/1.2/spec.html#id2774608).

RB_ID=680926