[util-core] `dedup` and `dedupWith` methods for Event
Problem
--

When registering callbacks on an `Event`, we often want to avoid
unnecessary work by ignoring updates that are not relevant. It is
often particularly useful to ignore update when the value did not
change.

Solution
--

Add two methods to `Event`: `dedup` and `dedupWith` to mask
updates. `dedup` filters updates that do not change the
underlying value (according to object equality), while
`dedupWith` allows to pass a custom predicate.

Result
--

We can register to an event stream without being notified of
events that are not of interest.

RB_ID=717437