util-stats: Add clear() to InMemoryStatsReceiver to remove all existing values
Motivation
InMemoryStatsReceiver is commonly used in tests. In cases where the StatsReceiver is
shared between tests, it is sometimes desirable to reset all the values in the StatsReceiver
to avoid cross-test interference.

Solution
Add .clear() method to InMemoryStatsReceiver which clears the internal data structures
used for collecting the stats.

Result
clear method added to InMemoryStatsReceiver

RB_ID=742481