Make Zookeeper -client and -server usages explicit across source repo.
Problem:
The current ZooKeeper jar is a monolithic artifact that contains client and
server classes. In order to make client and server usages clearer and prep
for a upcoming split of the artifact, we need to do some classification.

Solution:
Create placeholder zookeeper-client and zookeeper-server targets. Make those
usages explicit in build targets.

Result:
Dependencies are more clear, and it will be easier to introduce split artifacts.

RB_ID=751175