finagle-core, util-logging: Introduce HasLogLevel trait
Motivation

More control over the log level used for exceptions is necessary.
While `Failure` provides for this, it is marked as `final` and it can
be difficult to shoehorn other existing exception classes into it.

Solution

Introduce a small interface, `HasLogLevel` with a single method
`logLevel` that can be mixed into other exceptions as necessary.

Result

`Failure` mixes it in as does
`c.t.f.mux.ClientDiscardedRequestException`.

RB_ID=762874