finagle-stats: Add a flag for loading stats filters from a file
Problem

The only way to load stats filters is via a CSV list passed on the command line.

Solution

Add a flag which accepts a path to a file and load the file as the list of stats filters. Combine both lists for a complete list.

Only a new flag is added; current behavior does not change.

RB_ID=764914