finagle-mux: Refactor ClientDispatcher
Problem

The mux ClientDispatcher was becoming unwieldy. It handles translating
Requests/Responses to mux messages, managing control messages/the
state of a session, and tag management. By teasing these apart we can
make it easier to understand and more extensible.

Solution

The dispatcher now is primarily responsible for outstanding messages and
tag management, and thus, is a Service[Int => Message, Message]. This allows
us to move Request/Response translation to a filter above it. Finally, session
management is moved into a separate class.

RB_ID=760326