util-core: AsyncStream.mapConcurrent
Problem
-------

Frequently, when processing a stream of items, the exact order of
processing is not important, and it is desirable or necessary for
performance reasons to process items concurrently. At the same time,
it is usually necessary to have some sort of limit on the number
of concurrently-executed actions.

Solution
--------

Add the method mapConcurrent that concurrently executes an action
over a portion of the stream, yielding a new stream of results as
they become available. It's possible to achieve similar results by
using an AsyncSemaphore in code that is consuming from the stream,
but this method provides a high-level API for invoking concurrent/lazy
processing and accessing the results. In particular, it's valuable
for the API to return an AsyncStream, allowing the results of the
action to be operated on with the same API.

RB_ID=760948