util-logging: Add a flag to enable class and method name inference
Problem
-------

com.twitter.logging.QueueingHandler breaks
com.twitter.logging.LogRecord's inference of class and method
names, so they cannot appear in log output. The default
twitter-server log format attempts to include the class name and
method name, but the default logging handler uses a
QueueingHandler, so it does not appear.

Solution
--------

Provide an option in QueueingHandler to force the inference to
happen before the LogRecord gets queued and a flag to allow that
option to be set when using the default TwitterServer logger. The
flag is disabled by default, to prevent unintended latency impact
from creating stack traces in latency-sensitive threads.

RB_ID=793207